Adjust scene_obj functab signature to use directory/db_i rather than rt_db_internal
This is necessary to take advantage of cached data properly - by the time we
have cracked the rt_db_internal, we have already had to hit the .g disk data
and incur the memory penalty of loading all of the geometry information.  For
large BoTs, we want to get things like bounding boxes and coarse approximation
meshes without ever having to handle the full BoT at all (beyond the initial
pass to cache the LoD data, of course.)  The LMDB cache lets us do that, but
only if our processing flow allows us to completely avoid the rt_db_internal.

The reason this is possible with the OBJ functab is the d_minor_type entry in
the struct directory.  While we typically do OBJ lookups using ip->idb_type (or
its expanded form ip->idb_minor_type) it is also possible to do lookups with
the directory pointer itself.  I.e. OBJ[dp->d_minor_type] will also let us
access the per-primitive methods the same way OBJ[ip->idb_minor_type] would.
We can thus take advantage of this to provide a per-type method that does not
necessarily require the caller to crack the rt_db_internal (although it will
under the hood if that ends up being necessary.)

This also means that the scene_obj method will need to be passed the data
container holding any specialized cache data info that is not part of the
struct directory entry.  Right now that's just a stub for the function
definitions, but eventually it will most likely be a variation of the ged
draw_update_data_t container under the hood.

It might actually make more sense to associate the caches with the db_i struct,
since that's what they actually correspond to - cached database instance data.
If we go that route then we don't need a separate public cache struct at all...
both ged_lod and the dbi state ged_cache should be subsumed under the db_i
version, and will be accessible to the scene_obj method as part of dbip. It has
the further appeal of allowing the librt database I/O routines to see to the
clearing (and potentially the regeneration) of the cache data directly - that
would avoid a whole host of possible "forgot to sync the cache" scenarios.

Also document explicitly in a few places the expectation that
dp->d_minor_type == ip->idb_minor_type.