Stash a demo of how to truncate float numbers
May come in handy when trying to do a tolerance-based hash of mesh data, since
for that to work (i.e. generate an identical hash for meshes that are only the
same within tolerance) we'd need to clamp the number we're hashing to the data
subset that is above the tolerance specified.