Figuring out the limits of BoT hashing
The necessity of tolerance clamping means we remain sensitive to "almost but
not quite the same" points ending up getting clamped in opposite directions.
That is in the nature of clamping - you have to have some decision criteria,
and given enough inputs there will always be some cases that fall right on the
line.

What this means in practice is that bg_trimesh_hash can't be a full substitute
for bg_trimesh_diff, but it does still do enough to be interesting.  It might
be helpful, for example, to use to check if a BoT's data changed after an edit
operation - it might bias in favor of "changed" in cases like rotation or
translation but could still confirm unchanged if the original data is intact.

In testing with PCA meshes on a large .g file, it was also able to identify
about 80% of the matches the full bg_trimesh_diff check identified via matching
hash values.  In principle we might be able to combine the non-matching hash
values with additional info from bg_trimesh_diff results to do some intersting
things, but for the moment we'll file that for future consideration.