Various updates to dbi2 logic
Starting to think through the implications of how the libdm/libbv
data would interact with this state, and who should do what.

In the original qged work, the libdm object drawing was sort of hacked together
quickly to get the basic 3D scene elements drawing.  Circling back to the
problem it feels like libdm shouldn't be in the business of doing too much of
the view rendering - when to draw 3D scene objects, faceplate visuals, etc
feels like something higher levels of the stack may want to manage.  The lower
level "draw this scene obj in this view" dm_draw_obj call feels like a good
match. If we add a callback for such a function next to the bview struct dmp
pointer we already have, that gives us what we need in a view to trigger a
scene object rendering without any additional inputs.

If we go that route, rather than having bview trying to manage view-specific
and scene objects (and making a crazy complicated mess out of its internal
storage) that becomes the problem of the higher level logic. bview justs define
objects and libdm just renders them in a view.  The DbiState and BViewState
containers then hold comman and view-specific objects, doing whatever they want
to as far as reusing objects for multiple views or defining per-view objects.
An individual view's rendering method will handle what to draw for a specific
view and in what order, and we should be able to vastly simplify the bview and
dm level data management.