Restore default value for mged_rel_tol
Plotting after solid edit for rpc and rhc needs this value to be set or the
"restored" wireframe ends up coarse until a zap and redraw is completed.

Previous default was removed in e157a9e0a6, but the replacement of mged_rel_tol
took place in dc2fbcaf6561 so that would have been where we lost the default
global initialization.