Get remainder of libged compiling (not working yet)
Some of the view commands - particularly obj and polygon, and probably
lod as well - will need redesign. More things are probably broken than
not, but this gets us compiling.

Next up is gsh.  Once *that* is building, enough pieces will be in place
to begin seeing what will be involved with getting the various bits
actually working.

Until gsh and the libged drawing tests are back up and running, there's
not much point in working on any of the higher level libraries - there
may be more design adjustments that come out of the lower level work, so
no need to update them repeatedly just to compile for the sake of
compiling.