Replace one-off lmdb libged cache with bu_cache
Long term, the idea is to hide the implementation details of the cache
mechanism from most of BRL-CAD's code, since the operations are basically
generic.  The exception will probably be things like the librt prep cache,
which need to store REALLY large amounts of data and most likely aren't a good
fit for the LMDB backend.

The LoD cache is the next logical step - it should either merge with the Dbi
cache or at least also leverage libbu's cache API, but not quite sure yet
what the API implications are there.