Start looking at using bu_cache for LoD
This is a little thorny, but in the long run it should be worth it to minimize
our exposure to the LMDB API.  Also, using a generic bu_cache may allow us to
combine the LIBGED drawing info caching and the LoD caching info into a single
per-database cache, which would be conceptually simpler if it proves practical
and sufficiently fast/scalable.

If we want the librt aware aspects of the cache init to stay there, we may
want to migrate more logic from libged to librt.