Add some basic testing of the libbu cache
In particular, testing what happens when we run up against the database limits.
It seems we were previously failing silently on write, so now we confirm that a
written value can be read before bu_cache_write reports success.

The bu_cache test does some int and dbl writing in a larger database, and once
that succeeds deliberately makes a smaller one and confirms the expected
behavior both up to and at the limit - i.e., we look for failure return codes
once we try to put in more than the database can hold.

Probably will add more testing of this, but the most basic pieces do seem to be
in place now.