Start looking at a threading scheme for cache generation
Discussing with Sean, a simple way to recognize when we need to give up and
restart if someone edits the .g file before we're done would be to check the
file timestamp.  Unfortuantely, it looks like easy portable timestamp checking
needs C++20.  For now make a note, and focus instead on setting up the actual
processing logic - until we have the necessary C++20 feature or we implement
the per-platform wrappers might be able to do something like set an edited flag
in the db_i internal structure.  That won't catch changes to the .g from
another process, but would handle the most common case of one program to a .g
file.