Thinking about adding cache updating to the librt db_io layer
Early beginnings, but has some similarly to how callbacks were added.
If we have the cache info visible at this layer, there's no reason to
try to make some other layer do the work after the fact - we should have
what we need to take care of it directly.  If we can use a semaphore to
interrupt a cache init to make the changes and clean up whatever needs
cleaning up to allow the process to proceed/restart, that will probably
be the right answer for everything except another process mucking with
the .g (for which we will still probably need the timestamp trick.)