Do a bit of obb verification, rough in functab calls
The oriented bounding box routine calculates extent vectors, but there is more
than one set of extent vectors which can describe the same arb.  That can
result in different extent vectors and arb point ordering for the same bounding
volume depending on the inputs.

Checking the obb of the source mesh against the obb calculated from the arb
points generated from the first obb, these differences were observed and we got
different extent vectors for what was expected to be the same bounding arb.
In that case we go ahead and check that the two different extents end up
generating the same set of 3D points.

Looks like only BoT has any oriented bbox callback, and that API is returning
the 8 arb points rather than the center and extent vectors.  Should probably
fix the functab signature to return the latter, but in the meantime we can
bound the obb points to get a center and vectors.  That doesn't guarantee the
points will be in the same order, but as long as the result is a valid input
for drawing routines it shouldn't matter.