Validate what was written to the cache against the .g
Make sure we can read back what was written.  Can do this completely for
attribute based values and bounding boxes, but LoD we're still only checking
that we can load and exercise it rather than any sort of correctness check.

One possibility for LoD would be to generate it in the test program directly
from the bot internal, but would still need some sort of LoD comparator program
since the math probably shouldn't be trusted to recalculate everything with
enough precision for a memcmp test.