Add a C++ translation of SsdeepNET
Want to test out the ssdeep fuzzy hashing algorithm to group similar .g files.
Of the available implementations, SsdeepNET has an MIT license but was
implemented in C#.  Tried out Github Copilot to see if it could produce a C++
version of the SsdeepNET code.  A basic test with the gdiff regression test
examples seems to produce reasonable results, so go ahead and stash the header.

Next up will be to figure out a good way to analyze a lot of files to identify
and report groupings.