Experiment with TLSH hashing
ssdeep properties may not be a good match for what we're wanting to do
here, so set up a version of the tlsh algorithm to test out.  Leaving
ssdeep for now in case it ends up being useful later.

May need to try out Levenshtein distance as well - current idea is to
hash all the bu_external data of the objects with bu_data_hash, assemble
a sorted set of all those unsigned long long values, and then treat the
arrays of those values as input sequences for a Levenshtein distance
calculation - that could let us get a sense of how many objects have
changed geometry between two files. We could also independently do
something similar instead of the hash-the-string-of-names trick we are
currently trying...

Another trick to try might be to get editing distance metrics for file
names and pre-bin similarly named files for closer study.