implement different method for determining concavity.
we look at every line segment pairing and if the midpoint of any line is outside the arb, then we know we have a concave surface.  this is more resilient to vertex ordering issues and avoids needing to wind the right direction since we also make sure all faces go outward from the centroid.  in testing, this general approach appears to be far more robust, and considerably faster than convex hull methods.