restore LIBRT_BOT_MINTIE to adhere to policy and avoid churn.
since we still want a setting, no benefit changing the name and no apparent harm if users have scripts that set the value (if we don't want them doing that, we should announce deprecation). in testing, prep was fairly invariant to value changes; and shot was mildly influenced linearly by very large increases if the bots were big enough.