document the remaining tails so we can hopefully tag later today.
did encounter some minor changes in the last round of testing with robert that need to be validated.  rt matches, but getting different thickness (e.g., as gqa or nirt see) on just a handful of bot.  need to determine if it's an improvement or regression.

there's also a vertible slew of commits to look through for publicly visible changes.  we should toss the commmit log into gpt and see if it can help.