backtrace tool reliance is entirely an implementation detail.
never intended to be public API that it's using lldb or gdb or any other mechanism under the hood, and definitely not complexity that seems worth exposing all the way to userland as settings, docs, etc.

since this interface was developed, there are now some 3rd party library options available that could be considered/investigated for achieving similar capability.