Get dbi.h out of the public libged header directory
Shouldn't have been in there (at least not in this form) to start with, since
it isn't public API.  dbi.h is an experiment to create a C++ data model of the
.g database suitable for driving a Qt model that can support things like tree
views, but in this form it cannot be suitable as a public API (among many other
issues it exposes C++ containers, and the final form of this may belong in
librt rather than libged anyway.)  The concern in the initial stages was
searching for an implmentation approach that could in theory be performant
enough to scale to enormous databases, rather than isolating containers, so it
needs to be reworked to preserve its performance properties while also being a
proper API.  In any case, shouldn't have this form of the logic in this
location.

Until a more suitable replacement is designed and implemented, libqtcad still
needs it to drive its data model of the database, but it's not exposed in the
public libqtcad API so we can just reference the internal libged header in the
libqtcad implementation files for now.