Add working version of libbu cache API
This gets the rough (and previously unworking) version of the libbu cache API
into a working state, while also adding tests to verify its functionality in
various use cases.

The intent is that this will replace raw usage of the LMDB API in various parts
of BRL-CAD's code.  Not introducing those changes in this commit, since there
are other design considerations at play in those refactorings, but this piece
is self-contained and cleans up the preliminiary cache code that was already
present.