Adjust the link-time flag being used to suppress the bitv[2] warning
Our bitv_t bits[2] member triggers warnings such as the following:

warning: writing 1 byte into a region of size 0 [-Wstringop-overflow=] *out++ &= *in++;

We deliberately over-allocate extra memory for bits so we have something to
access, but per C99 this is undefined behavior: array subscripting a[b] is
defined as *(a + b) (C99 6.5.2.1p1) and pointer addition is only defined within
the same array object or one past its last element (C99 6.5.6p8). Accessing
elements bits[2]..bits[n-1] therefore goes outside the declared array object
and is UB (cataloged in Annex J.2: “array subscript is out of range” / “pointer
outside bounds of an array object”).

The current approach was used before C99 added flexible array member support
(and is still working, which is why we haven't changed it) but in theory a
sufficiently strict standards-compliant environment might someday break it
(hence the compiler warnings.) Since the "proper" fix would be to use a C99
flexible array member and fall back to this approach only in C++ compiling mode
(which only has flexible array members as extensions), we have settled for
suppressing the warning.  However, when we do LTO compilation the link
optimizer triggers another bounds analysis pass that doesn't respect other
warning suppressions we have defined.

We were avoiding a Werror compile failure with the previous compile flag, but
that still left the noisy warning messages in an LTO build.  This switchs us to
a full-on disabling of the warning during the link stage, which avoids printing
the messages altogether.

Now that we're using newer C standards, it might be worth revisiting whether
the bits member array could/should use the C99 approach...