Switch librt to using mmesh instead of gct_decimation
With the new version we need to supply stride in bytes, and optimization
needs sizeof rather than MD_FORMAT_INT.

We will be using the CONFIG style find_package for mmesh.

Compared to the old gct code, mmesh interpretes feature size a bit differently.
The old code's sensitivity to feature size was a fourth power formula, while
the new code uses a sixth power. The BRL-CAD code tries to adjust for this, but
there may also be other internal changes and the decimation results will most
likely still be a bit different.  Overall the tendency will be to produce finer
meshes for the same feature size, which will preserve detail better but also
result in more triangles as compared to the original settings.