Make edit state a separately allocated struct
Hmm.  Re-applying the MGED changes is proving a bit tricky - somewhere in the
refactor an editing state was mishandled, manifesting in an incorrect
appliation of tra logic (at least) when repeated edits are made in MGED.  To
help try and zero in on the issue, need to reapply safe changes to MGED main to
help minimize the differences between the working and non-working states (as
well as getting at least some of the refactoring reapplied.)

Note:  to reproduce the issue, the following works with moss.g from the MGED
cmd prompt:

facetize LIGHT light.bot
e light.bot
sed light.bot
tra 10 10 10
reject
sed light.bot
tra 10 10 10

The second tra in the sequence doesn't produce the same result as the first.