Restore MGED altered for librt editing in branch
Conceptually reverts commit 71776518c5, but it has a variety of changes
from that commit to adapt to more recent changes in MGED.

Add some thoughts about edit obj drawing, but probably need to shelve this for
now - consolidating the PARAM and MATRIX edit ops is very appealing, but the
undoubtedly tricky changes to the various drawing systems needed will be a
distraction from the edit refactor and unit testing.  We CAN make things work
as-is, so should do that first and then circle back to the bits that (badly)
need a rethink.