Try conditional execution of more complex tests
Per suggestion from Sean, see if we can make our CI tests a little more
measured - only try the other platforms if the Linux build succeeds,
and only try the static analyzer and code coverage if all platforms are
building and running.  This will take longer than executing everything
in parallel, so we won't get a "fully clean" result as quickly, but
the tradeoff is we don't waste a lot of time testing known broken
configurations on multiple platforms.