Round two - dependent execution of CI tests
This update to the GitHub CI yml files will attempt only the basic Ubuntu GCC
and Windows Ninja builds out of the gate.  If they succeed, other tests
dependent on those tests are executed - this avoids firing off a lot of extra
work if our base configurations are broken.

We also introduce caching of bext dependencies here.  It is MUCH more typical
for our bext to be stable while the main BRL-CAD repository is changing
quickly, so there is no need to rebuild bext every time we want to test the
main repo.  Previously we hadn't worked out out to use the GitHub mechanisms
for this, but with some Copilot AI assistance we now have a working setup.  If
bext changes it will still be necessary for the first BRL-CAD push that
encounters it to rebuild, but if we have a stable bext the actual BRL-CAD tests
should be considerably faster since we can now reuse previously built bext
binaries.

We also introduce a once-per-night run of the full bext-during-configure
integration.  BRL-CAD's configure needs to be able to build bext itself at need,
but the above optmiization actively avoids triggering that logic path.  Since
that test is more expensive, but we still want to know fairly quickly when it
breaks, we go with a once-a-day test of that mode on all platforms.

We have also moved the static analyzer and lcov builds to the once-per-day
mode - the static analyzer in particular is a relatively expensive build, and
we can generally spot fairly quickly from its output where issues would have
been introduced in a relatively small changeset.