Set up GitHub CI logic to run distcheck targets
The full distcheck target is too big to run as a single job within the
GitHub CI confines, but it turns out there's another way to get at it.
We already define individual build targets for each configuration, and
each individual target fits within the 6 hour limit for a single step.

Accordingly, if we define separate jobs for each distcheck target in the
yaml file, rather than trying to use the top level CMake distcheck build
target, we can avoid hitting the per-job limit.  The overall limit for
the entire process is longer, so even running in series we can stay
under both limits as long as we do one target at a time.

The catch is we don't want to hard code the list of distcheck targets in
distcheck.yml - that's an invitation for the .yml file to get stale and
out of sync as more configurations are added to distcheck.   The
workaround is to have BRL-CAD's CMake logic that sets up distcheck
targets also prepare an autogenerated distcheck.yml file, and then do a
comparison with the live copy in .github/workflows/distcheck.yml If they
are out of sync, the configure halts and instructs the developer to sync
the copies.

We deliberately don't write the new distcheck.yml directly into the
source tree.  This is the same pattern as the INSTALL file, which also
gets updated automatically, and the decision there as well was to
respect the "source tree is not altered by build logic" rule even though
it requires an extra manual step to do the updating.  The distcheck
target list does not change often, and this will catch the potential
syncing issue on those rare occasions when it does.