Be more selective about when we use Win32 console
For some Windows terminal environments like git bash (or VT aware consoles), we
want to handle things the way we would for other terminals.  If that doesn't work,
we need to fall back on the old-school console code path.

We want to have applications work in all those enviornments, so check up front
to detect what we're dealing with and set up our I/O handling accordingly.