Add brledit console editor
Originally based on https://github.com/troglobit/mg

We went with pico/nano style key bindings rather than emacs-style, since the
latter is more complex to deal with for a casual user.  Anyone wanting more
powerful editing capabilities is also likely to have their preferred editor
set up - the envisioned use case for this is constrained/last-resort cases
where the only option is to use what is available.

The original code has been extensively reworked, simplified, etc. to produce a
minimalist console editor that can work both on Linux/OSX/etc. type systems and
the various Windows console environments.  The latter in particular are very
difficult to deal with, and there will likely be various corner cases we
don't/can't handle well, but at least in early testing brledit is able to
functionally edit a file in both PowerShell and git bash.

The only environment where we might not have access to at least a vi editor is
a headless Windows server (which is why we included yedit in the first place)
or forcing a program not to use GUI editors on a Windows desktop without the
more standard 3rd party console Windows editors installed.  In that sense
yori's yedit did address that lack - git bash, for example, will generally
have nano available and so not need brledit - but brledit is both a much
smaller code footprint than yori and DOES work on POSIX type systems in case
some unforseen circumstance ends up making that desirable or necessary.

In addition to brledit itself we have added new logic to the libtermio.h header
to deal with the requirements of its interace, which are much more extensive
than any of the other BRL-CAD utilities still leveraging libtermio.h  For the
other apps compatibility wrappers were defined, and testing to make sure the
others are still working properly is one of the things we need to do before
merging this into main.