Add the SPSR meshing methodology to the pnts cmd
Previously this was only exposed indirectly via the facetize command, where it
was combined with decimation to be a fallback method of turning primitives into
triangles when MGED doesn't work.  However, SPSR itself actually takes point
clouds with normals, which is what the pnts command generates, so to allow for
easier experimenation and more control we can expose the method directly for
arbitrary point clouds: "pnts tri spsr input.pnts output.bot"

This is an outgrowth of experimenting with the Ball Pivot and Power Crust
methodologies to see if they would be useful, but neither is looking promising
in early testing so only the SPSR exposure and refactor to support bu_cmdtab
addition of other methods via new subcommands has been migrated back to main.

Ball pivot can produce meshes in some cases but they're not very good and
may need a lot of fiddling with the radius parameter - not to mention it seems
to be pretty easy to specify parameters that cause very long run times.

Power Crust I didn't explore as much, but it doesn't seem to be very well
suited to the arbitrarily complex shapes we might be throwing at it from
CAD geometry so I doubt it would provide us anything SPSR doesn't already.

Main reason to merge this in is to make it easier to try out new methodologies
currently being developed - Sean indicated there are some new approaches to
this problem that have been developed in the last couple years so this should
give us a convenient way to kick the tires for our use cases.