Looking into refactor of DbiState and new drawing layer
One of the complexities of the new state management used for qged's drawing
path is the large number of maps and hash keys.  If the memory overhead and
performance of using class containers to hold most of this instead is
reasonable, it should simplify some things - however, we need to test out the
performance.  Since the change is very invasive and will need to be implemented
in stages, set up a branch to hold the incremental work.

Looking into figuring out how to make the LoD setup a background process
triggered by db_open and cleaned up behind the scenes by db_close.  The
init_completed variable should also give us a way to avoid trying to pull data
from the cache before setup is complete.

Because this pulls in multithreading and disk I/O, it is hidden behind an
environment variable and not triggered by default.  An app can use bu_setenv
(either by default or reads from user settings) to stage the desired value
before calling any ged_open calls.

Also rethinking through the implications of how the libdm/libbv
data would interact with this state, and who should do what.

In the original qged work, the libdm object drawing was sort of hacked together
quickly to get the basic 3D scene elements drawing.  Circling back to the
problem it feels like libdm shouldn't be in the business of doing too much of
the view rendering - when to draw 3D scene objects, faceplate visuals, etc
feels like something higher levels of the stack may want to manage.  The lower
level "draw this scene obj in this view" dm_draw_obj call feels like a good
match. If we add a callback for such a function next to the bview struct dmp
pointer we already have, that gives us what we need in a view to trigger a
scene object rendering without any additional inputs.

If we go that route, rather than having bview trying to manage view-specific
and scene objects (and making a crazy complicated mess out of its internal
storage) that becomes the problem of the higher level logic. bview justs define
objects and libdm just renders them in a view.  The DbiState and BViewState
containers then hold comman and view-specific objects, doing whatever they want
to as far as reusing objects for multiple views or defining per-view objects.
An individual view's rendering method will handle what to draw for a specific
view and in what order, and we should be able to vastly simplify the bview and
dm level data management.

Rough in code to try and get a bu_gettime compatible file timestamp
Discussing with Sean, a simple way to recognize when we need to give up and
restart if someone edits the .g file before we're done would be to check the
file timestamp.  Unfortuantely, it looks like easy portable timestamp checking
needs C++20.  For now make a note, and focus instead on setting up the actual
processing logic - until we have the necessary C++20 feature or we implement
the per-platform wrappers might be able to do something like set an edited flag
in the db_i internal structure.  That won't catch changes to the .g from
another process, but would handle the most common case of one program to a .g
file.

The oriented bounding box routine calculates extent vectors, but there is more
than one set of extent vectors which can describe the same arb.  That can
result in different extent vectors and arb point ordering for the same bounding
volume depending on the inputs.

Checking the obb of the source mesh against the obb calculated from the arb
points generated from the first obb, these differences were observed and we got
different extent vectors for what was expected to be the same bounding arb.
In that case we go ahead and check that the two different extents end up
generating the same set of 3D points.

Looks like only BoT has any oriented bbox callback, and that API is returning
the 8 arb points rather than the center and extent vectors.  Should probably
fix the functab signature to return the latter, but in the meantime we can
bound the obb points to get a center and vectors.  That doesn't guarantee the
points will be in the same order, but as long as the result is a valid input
for drawing routines it shouldn't matter.