Improve performance of nmg_mdl_to_bot for all triangle case
Testing new DSP tess code, I observed a successful conversion that was spending
virtually all of its time (a LOT of it - order of minutes for what should have
been a couple seconds) in the nmg_mdl_to_bot code.  That's nonsensical, since
there's no real work to do (the data is already all triangles to begin with)
but it turns out that the nmg routine (which is probably a lot slower than it
shoudl be anyway) has REALLY serious scaling problems when presented with
something like a high-triangle count terrain model.

Rather than work around the problem (again), this time we actually introduce
into the nmg code itself a fast path that recognizes when we have the trivial
case and directly writes the data without further processing.  It doesn't fix
what's likely a host of other issues with nmg, but at least in this (extremely
common) case we should get tolerable performance now.