Get DSP NMG tessellation method working
Existing DSP tessellation via the librt rt_db_tess callback was
impractically slow, and the facetize command had been using the CM
fallback method (marching tetrahedrons) as the next best viable way to
produce a DSP mesh output.

This fix implements a basic meshing routine for digital elevation map
data that will complete within a reasonable timeframe - under 2 seconds
for the example terra.g file's terra.s primitive.  It does not
(currently) implement any of the more sophisticated terrain meshing
algorithms for the surface mesh, but it DOES go to some trouble to
generate a "bottom" mesh that closes the volume which is sparser than
the naive "every active cell gets 2 triangles" method.

This is not the more sophisticated terrain meshing I had originally
hoped to achieve (although that may still come one day.)  However, in
the meantime it *is* still a user visible improvement working right now
so I'll go ahead and put it in.