Update and enhance GED graph cmd
It's been a long time since this was tried, and Adaptagrams has apparently
changed quite a bit.  This update attempts to get things running with a newer
version of Adaptagrams (assuming a new CMake build system's outputs for
find_package).

We also add a basic SVG output mode (which is the default now) that uses higher
level Adaptagrams layouts to explore more visualization options.  This output
doesn't need the Tk window - any SVG viewer should be able to see the outputs,
and they do display in Firefox.

The actual motivation for touching this right now is to try and get a GED
plugin that depends on an LGPL library (and isn't centrally important
functionality) working to serve as an "always build as a plugin" test case.
The CMake build of Adaptagrams is a prelude to getting it running on Windows,
since any plugin adjustments will need to be tested there, but that will also
set the stage for making the graph command available reliably there as well
and lay the foundation for doing more interesting things with this capability.