Need *full* dependency setup for the source archives check
By bad luck, we had a cache miss in the bext CI dependencies due to an upstream
commit just before a Source Archives check fired, and it fell to Source
Archives to produce a new bext.  Problem was, the Qt dependencies weren't
present, so it couldn't do so successfully - build failed.

The way configure is set up, we *need* bext to be able to proceed, even with
just a Source Archives test - and changing that would be very difficult/
invasive for such a minor convience enhancment - we can already do the check,
as long as bext is set up correctly, and we need that for basically all other
cases anyway.  Better to just make sure we have the deps uniformly, so this can
succeed if we hit this corner case again.