Suppress the unused-parameter warning in lemon generated code
The error message code tried to work around this by using the variable
in question, but that is proving to be unreliable across different
versions of lemon - using yyminor appears to intrude on data structures
that lemon treats as internal, and what works with one version won't
work with another.  Instead, just use the pragmas to avoid the
compilation failure.