another stab at better scaling gqa grid
We need to always work in mm. Otherwise we restrict our grid to 1 UNIT -
ie if we're working in m, the grid can only go down to 1m = 1000mm (is a
very loose grid).

Also, set the grid lower bound to be 1/10th the upper to see if we can
get some cheap convergence and higher accuracy value without sacrificing
too much speed.
NOTE: this value is still largely arbitrary and would be nice to verify
how much error this is introducing over a much slower precision run.